% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculatePosition.R
\name{calculatePositionFragment}
\alias{calculatePositionFragment}
\title{CalculatePos}
\usage{
calculatePositionFragment(count_dir, alignment_path, gffPath, category)
}
\arguments{
\item{count_dir}{Output path directory}

\item{alignment_path}{SAM/BAM files with all alignment positiions}

\item{gffPath}{GTF/GFF path}

\item{category}{Category to calculate position.}
}
\description{
Function that calculates the position of the founded fragments in sequencing.
}
\author{
Alberto Rodriguez-Izquierdo
}
