% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfigFile.R
\name{getConfigFile}
\alias{getConfigFile}
\title{ConfigFileGet}
\usage{
getConfigFile(root)
}
\arguments{
\item{root}{Root path}
}
\value{
configFile
}
\description{
QC of all data
}
\author{
Alberto Rodriguez-Izquierdo, 2021
}
