% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictionSimulation.R
\name{graph_generator}
\alias{graph_generator}
\title{graph_gen}
\usage{
graph_generator(
  sequences,
  name_enzyme,
  min.size,
  max.size,
  configFile,
  root,
  nrepeat = NULL,
  use_replicate = NULL
)
}
\arguments{
\item{sequences}{Sequences to be graphed}

\item{name_enzyme}{name of the enzyme}

\item{min.size}{min size to plot}

\item{max.size}{max size to plot}

\item{configFile}{configFile}

\item{nrepeat}{number of repetitions}

\item{use_replicate}{if use replicate}
}
\description{
Graph generation from in-silico digestion.
}
\author{
Rodriguez-Izquierdo, Alberto, 2021
}
