% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfigFile.R
\name{nodesValidation}
\alias{nodesValidation}
\title{ConfigValidationsNodes}
\usage{
nodesValidation(configFile)
}
\arguments{
\item{configFile}{configFile}
}
\value{
configFile
}
\description{
ConfigValidationsNodes
}
\author{
Alberto Rodriguez-Izquierdo, 2021
}
