% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outputGeneration}
\alias{outputGeneration}
\title{Output}
\usage{
outputGeneration(results, root)
}
\arguments{
\item{results}{Input of results}

\item{root}{Working directory}

\item{configFile}{configFile}
}
\description{
Designing output
}
\author{
Alberto Rodriguez-Izquierdo, 2021
}
