% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomGenomeFragmentation.R
\name{randomGenomeFragmentation}
\alias{randomGenomeFragmentation}
\title{RandomGenomicFrag}
\usage{
randomGenomeFragmentation(genome_path, max_size, min_size, nb_repeats)
}
\arguments{
\item{genome_path}{Genome path in fasta format}

\item{max_size}{max size fragments}

\item{min_size}{min size fragments}

\item{nb_repeats}{number of repeats for random fragmentation.}
}
\value{
fragmentRes
}
\description{
Generation of random fragmentation from reference genome
}
\author{
Alberto Rodriguez-Izquierdo, 2021
}
