% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfigFile.R
\name{readConfigFile}
\alias{readConfigFile}
\title{ConfigReading}
\usage{
readConfigFile(root)
}
\arguments{
\item{root}{working directory}
}
\value{
configFile
}
\description{
App to start geoapp
}
\author{
Alberto Rodriguez-Izquierdo, 2021
}
