% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restrictionSimulation.R
\name{restrictionSimulation}
\alias{restrictionSimulation}
\title{SimulationRE}
\usage{
restrictionSimulation(
  dnaseq,
  enzyme_db,
  min.size,
  max.size,
  type_analysis,
  enzyme_selection = NULL,
  nb_repeat = NULL,
  use_output = NULL,
  outputDir = NULL,
  wd = NULL
)
}
\arguments{
\item{dnaseq}{File genomic data}

\item{enzyme_db}{Enzyme DB}

\item{min.size}{min size of length fragment}

\item{max.size}{max size of length fragment}

\item{type_analysis}{definition of analysis}

\item{enzyme_selection}{if selection enzymes, specify}

\item{nb_repeat}{number of repetitions}

\item{use_output}{Create output optional}

\item{outputDir}{output path}
}
\value{
results
}
\description{
Performance of in-silico restriction simulation from reference genome
}
\author{
Alberto Rodriguez-Izquierdo, 2021
}
