% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validations.R
\name{ValidateCharacter}
\alias{ValidateCharacter}
\alias{validateCharacter}
\title{validations Char}
\usage{
validateCharacter(configFile)
}
\arguments{
\item{configFile}{configFile}
}
\value{
configFile
}
\description{
App to start geoapp
}
\author{
Alberto Rodríguez Izquierdo, 2021
}
