% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfigFile.R
\name{validateConfigNodes}
\alias{validateConfigNodes}
\title{ConfigValidations}
\usage{
validateConfigNodes(nodes, configFile)
}
\arguments{
\item{configFile}{configFile}
}
\value{
configFile
}
\description{
App to start geoapp
}
\author{
Alberto Rodríguez Izquierdo, 2021
}
